requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-crispy-forms==1.11.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
ecdsa==0.14.1
idna==2.10
oauthlib==3.1.0
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
python-jose==3.2.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
